[lazy] refactor lazy init folder structure
