+++ tests/test_utils/test_lazy_init/lazy_init_utils.py
@@ -1,12 +1,13 @@
 import random
+from copy import deepcopy
 from typing import Any, Callable, Optional, Tuple
 
 import numpy as np
 import torch
 from packaging import version
 
+from colossalai.lazy.lazy_init import LazyInitContext, LazyTensor, _MyTensor
 from colossalai.tensor.d_tensor.layout_converter import to_global
-from colossalai.utils.model.experimental import LazyInitContext, LazyTensor, _MyTensor
 from tests.kit.model_zoo.registry import ModelAttribute
 
 SUPPORT_LAZY = version.parse(torch.__version__) >= version.parse('1.12.0')
@@ -65,8 +66,11 @@ def check_lazy_init(entry: TestingEntry, seed: int = 42, verbose: bool = False, 
     ctx = LazyInitContext()
     with ctx:
         deferred_model = model_fn()
+        # copied_deferred_model = deepcopy(deferred_model)
     deferred_model = ctx.materialize(deferred_model, verbose=verbose)
+    # copied_deferred_model = ctx.materialize(copied_deferred_model, verbose=verbose)
     assert_model_equal(model, deferred_model)
+    # assert_model_equal(deferred_model, copied_deferred_model)
     if check_forward:
         assert_forward_equal(model, deferred_model, data_gen_fn, output_transform_fn)
     if verbose:
